+++ src/move.cpp
@@ -46,7 +46,7 @@ OrderedMove::OrderedMove(uint8_t from, uint8_t to, uint8_t move_flag) : Move(fro
     this->value = 0;
 }
 
-OrderedMove::OrderedMove(uint8_t from, uint8_t to, uint8_t move_flag, uint16_t value) : Move(from, to, move_flag)
+OrderedMove::OrderedMove(uint8_t from, uint8_t to, uint8_t move_flag, int64_t value) : Move(from, to, move_flag)
 {
     this->value = value;
 }
